requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
idna==3.2
Pillow==8.3.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.6
